Unfied the flink file appender factory.
